Move test for TimeTrackedModel to the jobs tests and update it accordingly.
